requirements.txt
boto3==1.36.26
gradio==5.17.1
gradio_client==1.7.1
pandas==2.2.3
pillow==11.1.0
